feat(application): añadir puertos de entrada/salida a la capa de aplicación

+++ src/main/java/com/bcncgroup/pricingservice/prices/domain/Price.java
@@ -9,7 +9,7 @@ public record Price(
         LocalDateTime endDate,
         Long priceList,
         Long productId,
-        Long priority,
+        Integer priority,
         BigDecimal price,
         String currency) {
